requirements.txt
websockets
scipy
transformers
langchains
ctranslate2